requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.30.0
jinja2==3.1.4
python-dotenv==1.0.1
lxml==5.2.2
playwright==1.44.0
